Some small cleanup - Move ScriptLoader from Plugins to Core - Move TicketedCollection to arcana as weak_table and remove BabylonNativeUtils - Rename BabylonNativeCMakeFunctions to CMakeExtensions
