+++ Core/ScriptLoader/CMakeLists.txt
@@ -4,13 +4,12 @@ set(SOURCES
 
 add_library(ScriptLoader ${SOURCES})
 
-target_include_directories(ScriptLoader PRIVATE "Include/Babylon")
-target_include_directories(ScriptLoader INTERFACE "Include")
+target_include_directories(ScriptLoader PUBLIC "Include")
 
 target_link_to_dependencies(ScriptLoader
     PUBLIC napi
     PRIVATE arcana
     PRIVATE UrlLib)
 
-set_property(TARGET ScriptLoader PROPERTY FOLDER Plugins)
+set_property(TARGET ScriptLoader PROPERTY FOLDER Core)
 source_group(TREE ${CMAKE_CURRENT_SOURCE_DIR} FILES ${SOURCES})
