--- Dependencies/BabylonNativeUtils/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_library(BabylonNativeUtils INTERFACE)
-
-target_include_directories(BabylonNativeUtils INTERFACE "Include")
