+++ Plugins/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_subdirectory(NativeWindow)
-add_subdirectory(ScriptLoader)
 
 # Add NativeEngine
 # Dependencies: NativeWindow
