+++ Plugins/NativeWindow/CMakeLists.txt
@@ -10,7 +10,7 @@ target_include_directories(NativeWindow
 
 target_link_to_dependencies(NativeWindow 
     PUBLIC napi
-    PRIVATE BabylonNativeUtils
+    PRIVATE arcana
     PRIVATE JsRuntime)
 
 set_property(TARGET NativeWindow PROPERTY FOLDER Plugins)
@@ -20,5 +20,4 @@ add_library(NativeWindowInternal INTERFACE)
 target_include_directories(NativeWindowInternal INTERFACE "Source")
 target_link_to_dependencies(NativeWindowInternal
     INTERFACE NativeWindow
-    INTERFACE BabylonNativeUtils
     INTERFACE JsRuntime)
