+++ Plugins/NativeWindow/Source/NativeWindow.cpp
@@ -44,7 +44,7 @@ namespace Babylon::Plugins::Internal
             m_width = newWidth;
             m_height = newHeight;
 
-            m_onResizeCallbacks.ApplyToAll([this](auto& callback) {
+            m_onResizeCallbacks.apply_to_all([this](auto& callback) {
                 callback(m_width, m_height);
             });
         }
@@ -52,7 +52,7 @@ namespace Babylon::Plugins::Internal
 
     NativeWindow::OnResizeCallbackTicket NativeWindow::AddOnResizeCallback(OnResizeCallback&& callback)
     {
-        return m_onResizeCallbacks.Insert(std::move(callback));
+        return m_onResizeCallbacks.insert(std::move(callback));
     }
 
     void* NativeWindow::GetWindowPtr() const
